requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentencepiece>=0.1.99

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3

# Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0

# Data & Utils
pandas>=2.0.0
pyyaml>=6.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Interface
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Development
pytest>=7.4.0





